Its starting to resemble my initial idea.
It can now diff two runs, show you the git source diffs as well as
the time differences. Still should look to show the source differences
in color. Lots of error handling to clean up as well. Also need to
add more commands, such as removing a run, and letting the git user know
which commit a run would refer to before starting.